[featureMatching] Extract filterMatchesByMin2DMotion()
