+++ src/aliceVision/matching/matchesFiltering.hpp
@@ -8,6 +8,7 @@
 
 #include <aliceVision/sfmData/SfMData.hpp>
 #include <aliceVision/feature/Regions.hpp>
+#include <aliceVision/feature/RegionsPerView.hpp>
 #include <aliceVision/feature/feature.hpp>
 #include <aliceVision/matching/IndMatch.hpp>
 
@@ -65,6 +66,10 @@ void matchesGridFiltering(const aliceVision::feature::Regions& lRegions,
                           const std::pair<std::size_t, std::size_t>& rImgSize,
                           const aliceVision::Pair& indexImagePair,
                           aliceVision::matching::IndMatches& outMatches, size_t gridSize = 3);
+
+void filterMatchesByMin2DMotion(PairwiseMatches& mapPutativesMatches,
+                                const feature::RegionsPerView& regionPerView,
+                                double minRequired2DMotion);
 
 } // namespace sfm
 } // namespace aliceVision
